+++ src/query.c
@@ -2362,22 +2362,22 @@ int bitset_column_next(
 {
     /* Precomputed single-bit test */
     static const uint64_t bitmask[64] = {
-    1lu << 0, 1lu << 1, 1lu << 2, 1lu << 3,
-    1lu << 4, 1lu << 5, 1lu << 6, 1lu << 7,
-    1lu << 8, 1lu << 9, 1lu << 10, 1lu << 11,
-    1lu << 12, 1lu << 13, 1lu << 14, 1lu << 15,
-    1lu << 16, 1lu << 17, 1lu << 18, 1lu << 19,
-    1lu << 20, 1lu << 21, 1lu << 22, 1lu << 23,
-    1lu << 24, 1lu << 25, 1lu << 26, 1lu << 27,  
-    1lu << 28, 1lu << 29, 1lu << 30, 1lu << 31,
-    1lu << 32, 1lu << 33, 1lu << 34, 1lu << 35,  
-    1lu << 36, 1lu << 37, 1lu << 38, 1lu << 39,
-    1lu << 40, 1lu << 41, 1lu << 42, 1lu << 43,
-    1lu << 44, 1lu << 45, 1lu << 46, 1lu << 47,  
-    1lu << 48, 1lu << 49, 1lu << 50, 1lu << 51,
-    1lu << 52, 1lu << 53, 1lu << 54, 1lu << 55,  
-    1lu << 56, 1lu << 57, 1lu << 58, 1lu << 59,
-    1lu << 60, 1lu << 61, 1lu << 62, 1lu << 63
+    (uint64_t)1 << 0, (uint64_t)1 << 1, (uint64_t)1 << 2, (uint64_t)1 << 3,
+    (uint64_t)1 << 4, (uint64_t)1 << 5, (uint64_t)1 << 6, (uint64_t)1 << 7,
+    (uint64_t)1 << 8, (uint64_t)1 << 9, (uint64_t)1 << 10, (uint64_t)1 << 11,
+    (uint64_t)1 << 12, (uint64_t)1 << 13, (uint64_t)1 << 14, (uint64_t)1 << 15,
+    (uint64_t)1 << 16, (uint64_t)1 << 17, (uint64_t)1 << 18, (uint64_t)1 << 19,
+    (uint64_t)1 << 20, (uint64_t)1 << 21, (uint64_t)1 << 22, (uint64_t)1 << 23,
+    (uint64_t)1 << 24, (uint64_t)1 << 25, (uint64_t)1 << 26, (uint64_t)1 << 27,  
+    (uint64_t)1 << 28, (uint64_t)1 << 29, (uint64_t)1 << 30, (uint64_t)1 << 31,
+    (uint64_t)1 << 32, (uint64_t)1 << 33, (uint64_t)1 << 34, (uint64_t)1 << 35,  
+    (uint64_t)1 << 36, (uint64_t)1 << 37, (uint64_t)1 << 38, (uint64_t)1 << 39,
+    (uint64_t)1 << 40, (uint64_t)1 << 41, (uint64_t)1 << 42, (uint64_t)1 << 43,
+    (uint64_t)1 << 44, (uint64_t)1 << 45, (uint64_t)1 << 46, (uint64_t)1 << 47,  
+    (uint64_t)1 << 48, (uint64_t)1 << 49, (uint64_t)1 << 50, (uint64_t)1 << 51,
+    (uint64_t)1 << 52, (uint64_t)1 << 53, (uint64_t)1 << 54, (uint64_t)1 << 55,  
+    (uint64_t)1 << 56, (uint64_t)1 << 57, (uint64_t)1 << 58, (uint64_t)1 << 59,
+    (uint64_t)1 << 60, (uint64_t)1 << 61, (uint64_t)1 << 62, (uint64_t)1 << 63
     };
 
     /* Precomputed test to verify if remainder of block is set (or not) */
